# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Pro Backend API
  version: 0.2.0
  description: |
    Identifiers for persisted tenant resources are globally unique across tenants.
    tenant_id scopes ownership, authorization, and row-level security.
security:
  - bearerAuth: []
paths:
...
  /v1/hr/positions/ehrms/sync:
    post:
      summary: Sync positions from configured eHRMS
      responses:
        "200":
          description: eHRMS position sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EHRMSPositionSyncDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/hr/employees/ehrms/sync:
    post:
      summary: Sync employees from configured eHRMS
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EHRMSEmployeeSyncRequest"
      responses:
        "200":
          description: eHRMS sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EHRMSEmployeeSyncDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/ehrms/sync-runs:
    get:
      summary: List eHRMS sync runs
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        "200":
          description: Paginated sync runs
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EHRMSSyncRunPageDataResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/ehrms/sync:
    post:
      summary: Start a tracked manual eHRMS sync
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StartEHRMSSyncRequest"}
      responses:
        "200":
          description: Completed tracked sync run
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EHRMSSyncRunDetailDataResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/ehrms/sync-runs/{id}:
    get:
      summary: Get an eHRMS sync run and its steps
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Sync run detail
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EHRMSSyncRunDetailDataResponse"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/ehrms/sync-runs/{id}/retry:
    post:
      summary: Retry a failed or partial eHRMS sync run
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RetryEHRMSSyncRunRequest"}
      responses:
        "200":
          description: Retried sync run detail
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EHRMSSyncRunDetailDataResponse"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/authz/check:
    post:
      summary: Check one authorization request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckRequest"
      responses:
        "200":
          description: Authorization decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckResultDataResponse"
  /v1/authz/batch-check:
    post:
      summary: Check multiple authorization requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCheckRequest"
      responses:
        "200":
          description: Batch authorization decisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCheckResultDataResponse"
  /v1/authz/explain:
    post:
      summary: Explain one authorization decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckRequest"
      responses:
        "200":
          description: Authorization explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthzExplainDataResponse"
  /v1/authz/simulate:
    post:
      summary: Simulate one authorization decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthzSimulationRequest"
      responses:
        "200":
          description: Authorization simulation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthzSimulationDataResponse"
...
  /v1/org/units/ehrms/sync:
    post:
      summary: Sync organization units from configured eHRMS
      responses:
        "200":
          description: eHRMS organization unit sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EHRMSOrgUnitSyncDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/attendance/ehrms/sync:
    post:
      summary: Sync attendance daily summaries from configured eHRMS
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EHRMSAttendanceSyncRequest"
      responses:
        "200":
          description: eHRMS attendance sync summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EHRMSAttendanceSyncDataResponse"
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthenticated"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token issued by the configured identity provider and mapped to a local tenant account.
...
  responses:
    ValidationError:
      description: Validation failed. Field errors include tab, field, code and message; import row errors include row_number.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: Authenticated tenant/account context is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Permission, menu, button, field, or data-scope access was denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found in the current tenant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error. Response includes trace_id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    CheckResultDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CheckResult"
    BatchCheckResultDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/BatchCheckResult"
    AuthzExplainDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthzExplainResponse"
    AuthzSimulationDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthzSimulationResponse"
    CheckRequest:
      type: object
      required: [action]
      properties:
        application_code: {type: string}
        resource_type: {type: string}
        resource_id: {type: string}
        action: {type: string}
        context:
          type: object
          additionalProperties: true
        resource: {type: string}
        target: {type: string}
        scope: {type: string}
        target_employee_id: {type: string}
        route_method: {type: string}
        route_path: {type: string}
    CheckResult:
      type: object
      required: [allowed, action]
      properties:
        allowed: {type: boolean}
        reason: {type: string}
        matched_by:
          type: array
          items: {type: string}
        matched_permissions:
          type: array
          items: {type: string}
        missing_permissions:
          type: array
          items: {type: string}
        permission_set_ids:
          type: array
          items: {type: string}
        scope: {type: string}
        effective_scope: {type: string}
        conditions:
          type: object
          additionalProperties: true
        field_policies:
          type: object
          additionalProperties:
            type: string
        assumed_role:
          $ref: "#/components/schemas/AssumedRoleDecision"
        permission_boundary:
          type: object
          additionalProperties: true
        risk_level: {type: string}
        resource: {type: string}
        application_code: {type: string}
        resource_type: {type: string}
        resource_id: {type: string}
        action: {type: string}
        target: {type: string}
    BatchCheckRequest:
      type: object
      required: [checks]
      properties:
        checks:
          type: array
          items:
            $ref: "#/components/schemas/CheckRequest"
    BatchCheckResult:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/CheckResult"
    AssumedRoleDecision:
      type: object
      required: [role_id]
      properties:
        session_id: {type: string}
        role_id: {type: string}
        name: {type: string}
    AuthzExplainResponse:
      type: object
      required: [decision, evaluated_grants, scope_derivation]
      properties:
        decision:
          $ref: "#/components/schemas/CheckResult"
        evaluated_grants:
          type: array
          items:
            $ref: "#/components/schemas/AuthzEvaluatedGrant"
        deny_sources:
          type: array
          items: {type: string}
        boundary_effects:
          type: array
          items:
            $ref: "#/components/schemas/AuthzBoundaryEffect"
        scope_derivation:
          $ref: "#/components/schemas/AuthzScopeDerivation"
    AuthzEvaluatedGrant:
      type: object
      required: [source, matched, excluded_by]
      properties:
        source: {type: string}
        source_id: {type: string}
        permission_set_id: {type: string}
        permission: {type: string}
        effect: {type: string, enum: [allow, deny]}
        matched: {type: boolean}
        scope: {type: string}
        excluded_by:
          type: string
          nullable: true
    AuthzBoundaryEffect:
      type: object
      required: [source, effect, matched]
      properties:
        source: {type: string}
        permission: {type: string}
        effect: {type: string}
        matched: {type: boolean}
        excluded_by: {type: string}
    AuthzScopeDerivation:
      type: object
      properties:
        normal: {type: string}
        assumed: {type: string}
        boundary: {type: string}
        final: {type: string}
        intersection_empty: {type: boolean}
    AuthzSimulationRequest:
      type: object
      required: [check]
      properties:
        account_id:
          type: string
          description: Optional target account to simulate. Defaults to the authenticated account.
        check:
          $ref: "#/components/schemas/CheckRequest"
        overrides:
          $ref: "#/components/schemas/AuthzSimulationOverrides"
    AuthzSimulationOverrides:
      type: object
      properties:
        add_user_groups:
          type: array
          items: {type: string}
        remove_user_groups:
          type: array
          items: {type: string}
        add_permission_sets:
          type: array
          items: {type: string}
        remove_permission_sets:
          type: array
          items: {type: string}
        assume_role_id: {type: string}
        permission_set_changes:
          type: array
          items:
            $ref: "#/components/schemas/AuthzPermissionSetChange"
    AuthzPermissionSetChange:
      type: object
      required: [permission_set_id]
      properties:
        permission_set_id: {type: string}
        add_permissions:
          type: array
          items: {type: string}
        remove_permissions:
          type: array
          items: {type: string}
    AuthzSimulationResponse:
      type: object
      required: [before, after, diff]
      properties:
        before:
          $ref: "#/components/schemas/CheckResult"
        after:
          $ref: "#/components/schemas/CheckResult"
        diff:
          $ref: "#/components/schemas/AuthzSimulationDiff"
    AuthzSimulationDiff:
      type: object
      required: [allowed_changed, before_allowed, after_allowed, scope_changed]
      properties:
        allowed_changed: {type: boolean}
        before_allowed: {type: boolean}
        after_allowed: {type: boolean}
        scope_changed: {type: boolean}
        before_scope: {type: string}
        after_scope: {type: string}
        added_matched_by:
          type: array
          items: {type: string}
        removed_matched_by:
          type: array
          items: {type: string}
        added_matched_permissions:
          type: array
          items: {type: string}
        removed_matched_permissions:
          type: array
          items: {type: string}
...
    EHRMSEmployeeSyncDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EHRMSEmployeeSyncResponse"
    EHRMSSyncRunPageDataResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items, total, page, page_size, sort]
          properties:
            items: {type: array, items: {$ref: "#/components/schemas/EHRMSSyncRun"}}
            total: {type: integer}
            page: {type: integer}
            page_size: {type: integer}
            sort: {type: string}
    EHRMSSyncRunDetailDataResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/EHRMSSyncRunDetail"}
    EHRMSOrgUnitSyncDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EHRMSOrgUnitSyncResponse"
    EHRMSPositionSyncDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EHRMSPositionSyncResponse"
...
    EHRMSAttendanceSyncDataResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/EHRMSAttendanceSyncResponse"
...
    EHRMSEmployeeSyncRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [create, update, upsert]
          default: upsert
          description: create rejects existing employee numbers, update rejects missing employee numbers, upsert creates or updates by employee number.
    EHRMSAttendanceSyncRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [create, update, upsert]
          default: upsert
          description: create rejects existing employee/date summaries, update rejects missing summaries, upsert creates or updates by employee/date.
        since:
          type: string
          format: date
          description: Optional YYYY-MM-DD lower bound; older upstream summary rows are skipped.
...
    EHRMSEmployeeSyncResponse:
      type: object
      properties:
        fetched: {type: integer}
        created: {type: integer}
        updated: {type: integer}
        skipped: {type: integer}
        failed: {type: integer}
        departments_upserted: {type: integer}
        positions_upserted: {type: integer}
        mode: {type: string, enum: [create, update, upsert]}
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchEmployeeResult"
        row_errors:
          type: array
          items:
            $ref: "#/components/schemas/SyncRowError"
    EHRMSOrgUnitSyncResponse:
      type: object
      properties:
        fetched: {type: integer}
        upserted: {type: integer}
    EHRMSPositionSyncResponse:
      type: object
      properties:
        fetched: {type: integer}
        upserted: {type: integer}
    EHRMSAttendanceSyncResponse:
      type: object
      properties:
        fetched: {type: integer}
        created: {type: integer}
        updated: {type: integer}
        skipped: {type: integer}
        failed: {type: integer}
        leave_balances_fetched: {type: integer}
        leave_balances_upserted: {type: integer}
        leave_balances_skipped: {type: integer}
        leave_balances_failed: {type: integer}
        leave_details_fetched: {type: integer}
        leave_details_created: {type: integer}
        leave_details_updated: {type: integer}
        leave_details_skipped: {type: integer}
        leave_details_failed: {type: integer}
        mode: {type: string, enum: [create, update, upsert]}
        since: {type: string, format: date}
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchEmployeeResult"
        row_errors:
          type: array
          items:
            $ref: "#/components/schemas/SyncRowError"
    RetryEHRMSSyncRunRequest:
      type: object
      properties:
        mode: {type: string, enum: [create, update, upsert]}
        since: {type: string, format: date}
    StartEHRMSSyncRequest:
      type: object
      properties:
        mode: {type: string, enum: [create, update, upsert]}
        since: {type: string, format: date}
        include_attendance: {type: boolean, default: false}
    EHRMSSyncRun:
      type: object
      required: [id, tenant_id, account_id, sync_type, trigger_type, status, mode, attempt, max_attempts, retryable, started_at, created_at, updated_at]
      properties:
        id: {type: string}
        tenant_id: {type: string}
        account_id: {type: string}
        sync_type: {type: string, enum: [pipeline, employees, attendance]}
        trigger_type: {type: string, enum: [scheduled, manual, retry, cli]}
        status: {type: string, enum: [running, succeeded, partial, failed, skipped]}
        current_step: {type: string}
        mode: {type: string, enum: [create, update, upsert]}
        since: {type: string, format: date}
        attempt: {type: integer}
        max_attempts: {type: integer}
        retry_of_run_id: {type: string}
        request_id: {type: string}
        trace_id: {type: string}
        error_code: {type: string}
        error_message: {type: string}
        retryable: {type: boolean}
        next_retry_at: {type: string, format: date-time}
        summary: {type: object, additionalProperties: true}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    EHRMSSyncRunStep:
      type: object
      required: [id, tenant_id, run_id, step, sequence, status, attempt, started_at]
      properties:
        id: {type: string}
        tenant_id: {type: string}
        run_id: {type: string}
        step: {type: string}
        sequence: {type: integer}
        status: {type: string, enum: [running, succeeded, partial, failed, skipped]}
        attempt: {type: integer}
        error_code: {type: string}
        error_message: {type: string}
        summary: {type: object, additionalProperties: true}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
    EHRMSSyncRunDetail:
      type: object
      required: [run, steps]
      properties:
        run: {$ref: "#/components/schemas/EHRMSSyncRun"}
        steps: {type: array, items: {$ref: "#/components/schemas/EHRMSSyncRunStep"}}
...
    BatchEmployeeResult:
      type: object
      properties:
        row_number: {type: integer}
        employee_id: {type: string}
        success: {type: boolean}
        action: {type: string}
        code: {type: string}
        message: {type: string}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        tab:
          type: string
          description: Employee modal tab or section key, for example basic_info or employment_info.
        field:
          type: string
          description: Request or spreadsheet field key that failed validation.
        code:
          type: string
          description: Stable machine-readable validation or field-policy code.
        message:
          type: string
    RowError:
      type: object
      required: [row_number, field_errors]
      properties:
        row_number:
          type: integer
          minimum: 1
          description: Spreadsheet or CSV row number, including the header row offset used by the import parser.
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    SyncRowError:
      type: object
      properties:
        row_number: {type: integer}
        field: {type: string}
        code: {type: string}
        message: {type: string}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              type: integer
              format: int32
              description: |
                Numeric application error code. Current ranges:
                1xxxx common platform, request and authentication errors;
                2xxxx IAM and authorization errors; 3xxxx people-domain and HR errors;
                4xxxx attendance errors; 5xxxx workflow errors; 6xxxx agent errors.
              enum:
                - 10000
                - 10001
                - 10002
                - 10003
                - 10004
                - 10005
                - 10006
                - 10007
                - 10008
                - 10009
                - 10030
                - 10031
                - 10032
                - 10033
                - 10034
                - 10035
                - 10036
                - 10037
                - 10050
                - 10060
                - 10070
                - 20040
                - 20041
                - 20042
                - 20043
                - 20044
                - 20045
                - 30010
                - 30011
                - 30012
                - 30013
                - 30014
                - 30015
                - 30016
                - 30017
                - 30018
                - 30030
                - 30031
                - 30040
                - 30041
                - 30042
            message: {type: string}
            reason_code:
              type: string
              enum:
                - account_inactive
                - sso_email_not_authorized
                - sso_email_unverified
                - company_inactive
                - google_login_failed
                - sso_identity_conflict
                - sso_email_ambiguous
                - permission_missing
                - menu_denied
                - button_denied
                - field_denied
                - data_scope_denied
            field_errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
            row_errors:
              type: array
              items:
                $ref: "#/components/schemas/RowError"
            trace_id: {type: string}